refactor : visitLog 별도 dto 패키지 생성, recommendOrder null값으로 초기화하기 위해 Integer로 변경
