+++ src/main/java/com/seowonn/mymap/service/Impl/VisitLogServiceImpl.java
@@ -4,32 +4,23 @@ import static com.seowonn.mymap.type.ErrorCode.ACCESS_DENIED;
 import static com.seowonn.mymap.type.ErrorCode.FILES_EXCEED;
 import static com.seowonn.mymap.type.ErrorCode.INCORRECT_REGION;
 import static com.seowonn.mymap.type.ErrorCode.MY_MAP_NOT_FOUND;
-import static com.seowonn.mymap.type.ErrorCode.NO_LIKES_CLICKED;
 import static com.seowonn.mymap.type.ErrorCode.REGION_NOT_FOUND;
-import static com.seowonn.mymap.type.ErrorCode.USER_NOT_FOUND;
 import static com.seowonn.mymap.type.ErrorCode.VISIT_LOG_NOT_FOUND;
 
-import com.seowonn.mymap.dto.NewVisitLogDto;
-import com.seowonn.mymap.dto.UpdateVisitLogDto;
+import com.seowonn.mymap.dto.visitLog.NewVisitLogDto;
+import com.seowonn.mymap.dto.visitLog.UpdateVisitLogDto;
 import com.seowonn.mymap.entity.Image;
-import com.seowonn.mymap.entity.Likes;
-import com.seowonn.mymap.entity.Member;
 import com.seowonn.mymap.entity.MyMap;
 import com.seowonn.mymap.entity.SiGunGu;
 import com.seowonn.mymap.entity.VisitLog;
 import com.seowonn.mymap.exception.MyMapSystemException;
-import com.seowonn.mymap.repository.LikesRepository;
-import com.seowonn.mymap.repository.MemberRepository;
 import com.seowonn.mymap.repository.MyMapRepository;
 import com.seowonn.mymap.repository.SiGunGuRepository;
 import com.seowonn.mymap.repository.VisitLogRepository;
 import com.seowonn.mymap.service.CheckService;
 import com.seowonn.mymap.service.VisitLogService;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
 import jakarta.transaction.Transactional;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
