+++ src/main/java/com/seowonn/mymap/service/VisitLogService.java
@@ -1,7 +1,7 @@
 package com.seowonn.mymap.service;
 
-import com.seowonn.mymap.dto.NewVisitLogDto;
-import com.seowonn.mymap.dto.UpdateVisitLogDto;
+import com.seowonn.mymap.dto.visitLog.NewVisitLogDto;
+import com.seowonn.mymap.dto.visitLog.UpdateVisitLogDto;
 import com.seowonn.mymap.entity.VisitLog;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
